rainbow_licorne/requirements.txt
django
django-filter
pillow
